move dependenciesutils
