+++ utils/dependencies/utils.go
@@ -1,4 +1,4 @@
-package utils
+package dependencies
 
 import (
 	"errors"
