+++ utils/dependencies/utils_test.go
@@ -1,4 +1,4 @@
-package utils
+package dependencies
 
 import (
 	"testing"
